[refactor] rename db connection and move to utils
